# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OSM2CAI",
        "description": "API documentation",
        "contact": {
            "name": "Contact the developers",
            "email": "catastorei@cai.it"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8008",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v2/ecpois/bb/{bounding_box}/{type}": {
            "get": {
                "tags": [
                    "Api V2"
                ],
                "summary": "Get Ec POIs by Bounding Box and Type",
                "description": "Returns a list of Eco POIs within the specified bounding box and of the specified type",
                "operationId": "5a4cd0905ca2a8c4f71fbc06be3eff16",
                "parameters": [
                    {
                        "name": "bounding_box",
                        "in": "path",
                        "description": "Bounding box in 'minLng,minLat,maxLng,maxLat' format",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Type of the POIs to retrieve: R for relation, W for way, N for node",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v2/ecpois/{hr_osm2cai_id}/{type}": {
            "get": {
                "tags": [
                    "Api V2"
                ],
                "summary": "Get EcPOIs in a 1km buffer from the HikingRoutes defined by ID",
                "description": "Returns a list of Ec POIs around 1km from a specific OSM2CAI hiking route ID and of a specified type",
                "operationId": "d169986961734438c0ac420747a64a46",
                "parameters": [
                    {
                        "name": "hr_osm2cai_id",
                        "in": "path",
                        "description": "OSM2CAI ID of the hiking route",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Type of the POIs to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v2/ecpois/osm/{hr_osm_id}/{type}": {
            "get": {
                "tags": [
                    "Api V2"
                ],
                "summary": "Get EcPOIs in a 1km buffer from the HikingRoutes defined by OSM ID",
                "description": "Returns a list of Ec POIs around 1km from a specific OSM hiking route ID and of a specified type",
                "operationId": "91f3f11f8731dc75060881b68ba065e1",
                "parameters": [
                    {
                        "name": "hr_osm_id",
                        "in": "path",
                        "description": "OSM ID of the hiking route",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Type of the POIs to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}